+++ microbit_v2/src/animations/carrousel.rs
@@ -1,5 +1,6 @@
 use crate::animations;
-use crate::animations::{Animation, Carrousel, NUM_COLORS, NUM_LEDS, Settings};
+use crate::animations::{Animation, Carrousel, NUM_COLORS, NUM_LEDS};
+use cookie_monster_common::animations::Settings;
 use core::cell::RefCell;
 use embedded_hal::delay::DelayNs;
 use microbit::hal::Timer;
@@ -26,14 +27,14 @@ impl<'a> Carrousel<'a> {
 
 impl Animation for Carrousel<'_> {
     fn brightness(&self, settings: &Settings) -> f32 {
-        settings.brightness * 0.05
+        settings.brightness() * 0.05
     }
 
     fn render(
         &mut self, ws2812: &mut Ws2812<Spi<SPI0>>, timer: &mut Timer<TIMER0>, settings: &Settings,
     ) {
         self.data.borrow_mut()[self.position] = animations::create_color_with_brightness(
-            &animations::COLORS[self.color_index],
+            animations::COLORS[self.color_index],
             self.brightness(settings),
         );
 
@@ -48,8 +49,8 @@ impl Animation for Carrousel<'_> {
             self.color_index = new_color;
         }
 
-        ws2812.write(self.data.borrow().iter().cloned()).unwrap();
-        timer.delay_ms(settings.delay);
+        ws2812.write(self.data.borrow().iter().copied()).unwrap();
+        timer.delay_ms(settings.delay());
     }
 
     fn reset(&mut self) {